fix: address PR 2826 review findings and rebase over landed #2819
- envelope reader skips properties subtrees (Codex P2) + regression test
- floating division in resolution guard byte cap; LINQ filters; test float literals
- port evaluation_no_leader gauge into the IMeterFactory metrics pattern
- reword sourceCrs prose that tripped the #2732 SRID-allowlist ratchet (trunk-inherited)
- regenerate feature catalog over the rebased assembly

+++ src/Honua.Worker.Gdal/Execution/GdalOutputGridGuard.cs
@@ -87,7 +87,7 @@ internal static class GdalOutputGridGuard
         // output size is pixels × 8. Compare via division to avoid an Int64 overflow on
         // the multiply when the pixel caps are configured very high.
         const long BytesPerFloat64Pixel = 8L;
-        if (pixels > options.MaxDecodedRasterBytes / BytesPerFloat64Pixel)
+        if (pixels > (double)options.MaxDecodedRasterBytes / BytesPerFloat64Pixel)
         {
             error = $"estimated output grid size {pixels.ToString(CultureInfo.InvariantCulture)} pixels × 8 bytes/pixel (single-band Float64) exceeds configured MaxDecodedRasterBytes={options.MaxDecodedRasterBytes.ToString(CultureInfo.InvariantCulture)}";
             return false;
